matrix/matrixserver/requirements.txt
datrie
django
djangorestframework
pandas